xtask: split corpus.rs into a directory module
Pure code move: mod.rs keeps fixture loading, with contract/runner/
summary/markdown/report as siblings, each under the 800-line shape
ceiling. check_contract and run_with_ceiling are split into smaller
functions in the process, so their too_many_lines ratchet allows are
removed. residue.rs's gate-reachability test now scans every corpus/
file instead of the one that no longer exists.
