+++ xtask/src/corpus/mod.rs
@@ -41,6 +41,18 @@ use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 
+mod contract;
+mod markdown;
+mod report;
+mod runner;
+mod summary;
+
+pub(crate) use contract::*;
+pub(crate) use markdown::*;
+pub(crate) use report::*;
+pub(crate) use runner::*;
+pub(crate) use summary::*;
+
 /// Mechanical net for the O(n²)-in-a-loop class of bug (AGENTS.md: a
 /// degenerate input once took 153s instead of milliseconds). Deliberately
 /// coarse — 100ms is nowhere near a single-fixture in-memory parse's real
@@ -99,7 +111,7 @@ struct ToolMeta {
 /// promises that may only weaken via an explicit, reviewable edit to this
 /// file. Every field is optional; an absent field asserts nothing.
 #[derive(Debug, Clone, Default, Deserialize)]
-struct ContractMeta {
+pub(crate) struct ContractMeta {
     #[serde(default)]
     expected_framework: Option<String>,
     #[serde(default)]
@@ -166,7 +178,7 @@ struct ContractMeta {
 /// cover a bless that included a full prose read (`corpus/README.md`).
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize)]
 #[serde(rename_all = "snake_case")]
-enum VerdictScope {
+pub(crate) enum VerdictScope {
     Flags,
     Subcommands,
     Descriptions,
@@ -238,7 +250,7 @@ struct BlessMeta {
 /// `Human` or `AgentThenHuman`.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize)]
 #[serde(rename_all = "kebab-case")]
-enum BlessProvenance {
+pub(crate) enum BlessProvenance {
     Human,
     AgentThenHuman,
     Agent,
@@ -493,1511 +505,6 @@ pub(crate) fn discover_fixtures(corpus_root: &Path) -> anyhow::Result<Vec<Fixtur
         }
     }
     Ok(out)
-}
-
-/// Contract-weakening detection: lowering `min_subcommands`, shrinking
-/// `must_contain_flags`, or marking a previously-enforced fixture
-/// `[xfail]` all make a real failure disappear silently.
-/// `corpus/README.md` permits weakening a contract via an explicit,
-/// justified edit — this makes sure a reviewer sees it happen.
-///
-/// This module has no git access (`tests/no_process_outside_exec.rs`
-/// forbids `std::process` in `xtask/src`), so it diffs `[contract]` fields
-/// between the current corpus and a second plain directory
-/// (`baseline_root`, populated by whatever invokes this binary via
-/// `--baseline-dir`, e.g. `git archive <base-ref> corpus | tar -x`). With
-/// no `--baseline-dir`, this function is never called.
-///
-/// Returns one `"CONTRACT WEAKENED: <fixture> <field>"` line per weakened
-/// field. Reported, not gated — a contract may legitimately weaken.
-fn contract_weakened_lines(current: &[Fixture], baseline: &[Fixture]) -> Vec<String> {
-    let mut lines = Vec::new();
-    for base in baseline {
-        let Some(now) = current.iter().find(|f| f.label == base.label) else {
-            lines.push(format!(
-                "CONTRACT WEAKENED: {} fixture-removed (present in baseline, missing now)",
-                base.label
-            ));
-            continue;
-        };
-        let (b, n) = (&base.meta.contract, &now.meta.contract);
-
-        // A framework assertion that's simply gone is a removed check —
-        // never flagged for merely *changing* to a different framework
-        // name, since that has no natural "weaker/stronger" ordering and a
-        // real detection improvement legitimately changes it.
-        if b.expected_framework.is_some() && n.expected_framework.is_none() {
-            lines.push(format!(
-                "CONTRACT WEAKENED: {} expected_framework (assertion removed)",
-                base.label
-            ));
-        }
-
-        if let Some(base_status) = &b.min_status {
-            let base_rank = crate::status::status_rank(base_status);
-            let now_rank = n.min_status.as_deref().and_then(crate::status::status_rank);
-            if now_rank < base_rank {
-                lines.push(format!(
-                    "CONTRACT WEAKENED: {} min_status ({:?} -> {:?})",
-                    base.label,
-                    base_status,
-                    n.min_status.as_deref().unwrap_or("(removed)"),
-                ));
-            }
-        }
-
-        if let Some(base_min) = b.min_subcommands {
-            let now_min = n.min_subcommands.unwrap_or(0);
-            if now_min < base_min {
-                lines.push(format!(
-                    "CONTRACT WEAKENED: {} min_subcommands ({base_min} -> {now_min})",
-                    base.label
-                ));
-            }
-        }
-
-        let missing_flags: Vec<&str> = b
-            .must_contain_flags
-            .iter()
-            .filter(|spec| !n.must_contain_flags.iter().any(|s| s == *spec))
-            .map(String::as_str)
-            .collect();
-        if !missing_flags.is_empty() {
-            lines.push(format!(
-                "CONTRACT WEAKENED: {} must_contain_flags (dropped: {})",
-                base.label,
-                missing_flags.join(", ")
-            ));
-        }
-
-        // A negative claim weakens by *losing an entry*, exactly as a
-        // positive one does — the direction of the claim flips, the
-        // direction of its weakening does not. Dropping
-        // `must_not_contain_flags = ["---...---"]` retires the only
-        // statement that the mariadb ruler is a phantom, and would let the
-        // defect return unremarked. Adding an entry tightens, and is never
-        // flagged, same as `must_contain_flags`.
-        let dropped_forbidden: Vec<&str> = b
-            .must_not_contain_flags
-            .iter()
-            .filter(|spec| !n.must_not_contain_flags.iter().any(|s| s == *spec))
-            .map(String::as_str)
-            .collect();
-        if !dropped_forbidden.is_empty() {
-            lines.push(format!(
-                "CONTRACT WEAKENED: {} must_not_contain_flags (dropped: {})",
-                base.label,
-                dropped_forbidden.join(", ")
-            ));
-        }
-
-        let missing_positionals: Vec<&str> = b
-            .must_contain_positionals
-            .iter()
-            .filter(|name| !n.must_contain_positionals.iter().any(|s| s == *name))
-            .map(String::as_str)
-            .collect();
-        if !missing_positionals.is_empty() {
-            lines.push(format!(
-                "CONTRACT WEAKENED: {} must_contain_positionals (dropped: {})",
-                base.label,
-                missing_positionals.join(", ")
-            ));
-        }
-
-        let missing_modifiers: Vec<&str> = b
-            .must_contain_modifiers
-            .iter()
-            .filter(|name| !n.must_contain_modifiers.iter().any(|s| s == *name))
-            .map(String::as_str)
-            .collect();
-        if !missing_modifiers.is_empty() {
-            lines.push(format!(
-                "CONTRACT WEAKENED: {} must_contain_modifiers (dropped: {})",
-                base.label,
-                missing_modifiers.join(", ")
-            ));
-        }
-
-        let missing_env_vars: Vec<&str> = b
-            .must_contain_env_vars
-            .iter()
-            .filter(|name| !n.must_contain_env_vars.iter().any(|s| s == *name))
-            .map(String::as_str)
-            .collect();
-        if !missing_env_vars.is_empty() {
-            lines.push(format!(
-                "CONTRACT WEAKENED: {} must_contain_env_vars (dropped: {})",
-                base.label,
-                missing_env_vars.join(", ")
-            ));
-        }
-
-        for (path, base_specs) in &b.must_contain_flags_by_path {
-            let now_specs = n.must_contain_flags_by_path.get(path);
-            let missing: Vec<&str> = base_specs
-                .iter()
-                .filter(|spec| !now_specs.is_some_and(|specs| specs.iter().any(|s| s == *spec)))
-                .map(String::as_str)
-                .collect();
-            if !missing.is_empty() {
-                lines.push(format!(
-                    "CONTRACT WEAKENED: {} must_contain_flags_by_path[{path:?}] (dropped: {})",
-                    base.label,
-                    missing.join(", ")
-                ));
-            }
-        }
-
-        let base_xfail = base.meta.xfail.as_ref().is_some_and(|x| x.broken);
-        let now_xfail = now.meta.xfail.as_ref().is_some_and(|x| x.broken);
-        if !base_xfail && now_xfail {
-            lines.push(format!(
-                "CONTRACT WEAKENED: {} xfail (newly marked broken — contract failures no longer fail the run)",
-                base.label
-            ));
-        }
-    }
-    lines
-}
-
-/// Extract a fixture's full tree: root extraction, then a bounded
-/// recursive fill into every discovered subcommand (see this module's doc
-/// comment). Returns `None` when no tier produced a root at all.
-pub(crate) fn extract_tree(runner: &Runner, resolved: &ResolvedTool) -> Option<CommandNode> {
-    let result = runner.extract_full_for(resolved);
-    let root = result.root?;
-    let mut budget = MAX_FIXTURE_NODES.saturating_sub(1);
-    Some(warm(
-        runner,
-        resolved,
-        root,
-        std::slice::from_ref(&resolved.name),
-        &mut budget,
-    ))
-}
-
-fn warm(
-    runner: &Runner,
-    resolved: &ResolvedTool,
-    mut node: CommandNode,
-    path: &[String],
-    budget: &mut usize,
-) -> CommandNode {
-    let children = std::mem::take(&mut node.subcommands);
-    let mut filled = Vec::with_capacity(children.len());
-    for child in children {
-        if *budget == 0 {
-            filled.push(child);
-            continue;
-        }
-        *budget -= 1;
-        let mut child_path = path.to_vec();
-        child_path.push(child.name.clone());
-        let fill = runner.fill_node(resolved, &child_path, child);
-        filled.push(warm(runner, resolved, fill.node, &child_path, budget));
-    }
-    node.subcommands = filled;
-    node
-}
-
-/// Render `node` in exactly the format `expected.snap` fixtures use:
-/// `mandible_core::to_snapshot` through plain `serde_yaml::to_string`,
-/// **not** `insta`'s snapshot macro. `insta` prepends a `source:`/
-/// `expression:` header meant for its own review workflow, which a CLI
-/// binary can't drive sanely (`insta`'s harness assumes a test binary and
-/// per-test dynamic snapshot paths) — see the corpus work order's design
-/// decision on why this crate does a plain file compare instead. Using
-/// the same `to_snapshot` + `serde_yaml` pair the format is defined by
-/// (`mandible_core::snapshot`'s doc comment) is what keeps this single-
-/// sourced rather than a second, driftable serialization.
-fn render_snapshot(node: &CommandNode) -> anyhow::Result<String> {
-    let snapshot = mandible_core::to_snapshot(node);
-    serde_yaml::to_string(&snapshot).map_err(|e| anyhow::anyhow!("serializing snapshot: {e}"))
-}
-
-/// The result of comparing a fixture's freshly-rendered tree against its
-/// `expected.snap`.
-enum SnapshotCheck {
-    /// No tier produced a root at all — there is nothing to compare.
-    NoRoot,
-    /// `expected.snap` doesn't exist. Legal only for an `[xfail]` fixture
-    /// (`corpus/README.md` step 4: "a fixture marked broken has no
-    /// expected tree yet").
-    Missing,
-    /// Byte-identical.
-    Match,
-    /// Differs, starting at the given 1-indexed line.
-    Mismatch {
-        line: usize,
-        expected: String,
-        actual: String,
-    },
-}
-
-fn check_snapshot(fixture: &Fixture, root: Option<&CommandNode>) -> anyhow::Result<SnapshotCheck> {
-    let Some(root) = root else {
-        return Ok(SnapshotCheck::NoRoot);
-    };
-    let snap_path = fixture.expected_snap_path();
-    if !snap_path.is_file() {
-        return Ok(SnapshotCheck::Missing);
-    }
-    let expected = std::fs::read_to_string(&snap_path)
-        .map_err(|e| anyhow::anyhow!("reading {}: {e}", snap_path.display()))?;
-    let actual = render_snapshot(root)?;
-    if expected == actual {
-        return Ok(SnapshotCheck::Match);
-    }
-    let expected_lines: Vec<&str> = expected.lines().collect();
-    let actual_lines: Vec<&str> = actual.lines().collect();
-    let first_diff = expected_lines
-        .iter()
-        .zip(actual_lines.iter())
-        .position(|(e, a)| e != a)
-        .unwrap_or_else(|| expected_lines.len().min(actual_lines.len()));
-    Ok(SnapshotCheck::Mismatch {
-        line: first_diff + 1,
-        expected: expected_lines
-            .get(first_diff)
-            .unwrap_or(&"<end of file>")
-            .to_string(),
-        actual: actual_lines
-            .get(first_diff)
-            .unwrap_or(&"<end of file>")
-            .to_string(),
-    })
-}
-
-/// A single `[contract]` field that failed, human-readable and naming the
-/// actual value alongside what was required — spec's own example of a
-/// good failure message (`corpus/README.md`'s companion work order):
-/// "git: min_subcommands 20, got 23 — OK; snapshot mismatch at
-/// .positionals[1].name".
-struct ContractFailure(String);
-
-/// Check every field the `[contract]` sets against `root`, returning one
-/// [`ContractFailure`] per violated field (empty = every check that was
-/// actually specified passed). A field left unset in `meta.toml` asserts
-/// nothing and is silently skipped.
-// Ratchet: one check per contract field, flat and deliberately unfactored. Listed in scripts/ratchet.txt.
-#[allow(clippy::too_many_lines)]
-fn check_contract(contract: &ContractMeta, root: Option<&CommandNode>) -> Vec<ContractFailure> {
-    let mut failures = Vec::new();
-    let Some(root) = root else {
-        // No root at all trivially fails every contract field that was
-        // actually specified — name them all rather than one opaque
-        // "no root" line, so the report reads the same shape whether the
-        // failure is "wrong tree" or "no tree".
-        if contract.expected_framework.is_some() {
-            failures.push(ContractFailure(
-                "expected_framework: no root produced".into(),
-            ));
-        }
-        if contract.min_status.is_some() {
-            failures.push(ContractFailure("min_status: no root produced".into()));
-        }
-        if contract.min_subcommands.is_some() {
-            failures.push(ContractFailure("min_subcommands: no root produced".into()));
-        }
-        if !contract.must_contain_flags.is_empty() {
-            failures.push(ContractFailure(
-                "must_contain_flags: no root produced".into(),
-            ));
-        }
-        if !contract.must_contain_flags_by_path.is_empty() {
-            failures.push(ContractFailure(
-                "must_contain_flags_by_path: no root produced".into(),
-            ));
-        }
-        if !contract.must_contain_positionals.is_empty() {
-            failures.push(ContractFailure(
-                "must_contain_positionals: no root produced".into(),
-            ));
-        }
-        if !contract.must_contain_modifiers.is_empty() {
-            failures.push(ContractFailure(
-                "must_contain_modifiers: no root produced".into(),
-            ));
-        }
-        if !contract.must_contain_env_vars.is_empty() {
-            failures.push(ContractFailure(
-                "must_contain_env_vars: no root produced".into(),
-            ));
-        }
-        // `must_not_contain_flags` is deliberately absent from this list.
-        // Every field above is a positive claim, which a missing tree
-        // trivially breaks — "the tool has --paginate" cannot hold of no
-        // tree. A negative claim is the opposite: "no root flag is spelled
-        // X" is *satisfied* by a tree with no flags at all, so reporting it
-        // here would announce a violation of a promise that in fact holds,
-        // which is a false positive in the one place this runner's
-        // authority comes from. A fixture that produced no root still fails
-        // loudly — on its snapshot, and on every positive field it set.
-        return failures;
-    };
-
-    if let Some(expected) = &contract.expected_framework {
-        let actual = root
-            .detected_framework
-            .clone()
-            .unwrap_or_else(|| "generic".to_string());
-        if &actual != expected {
-            failures.push(ContractFailure(format!(
-                "expected_framework: expected {expected:?}, got {actual:?}"
-            )));
-        }
-    }
-
-    if let Some(min_status) = &contract.min_status {
-        let result_stub = extraction_result_stub(root.clone());
-        let status = crate::status::compute(&result_stub);
-        if !crate::status::meets_min_status(status.label, min_status) {
-            failures.push(ContractFailure(format!(
-                "min_status: required at least {min_status:?}, got {:?}",
-                status.label
-            )));
-        }
-    }
-
-    if let Some(min) = contract.min_subcommands {
-        let got = root.subcommands.len();
-        if got < min {
-            failures.push(ContractFailure(format!(
-                "min_subcommands: required at least {min}, got {got}"
-            )));
-        }
-    }
-
-    let missing_flags: Vec<&str> = contract
-        .must_contain_flags
-        .iter()
-        .filter(|spec| !flag_present(root, spec))
-        .map(|s| s.as_str())
-        .collect();
-    if !missing_flags.is_empty() {
-        failures.push(ContractFailure(format!(
-            "must_contain_flags: missing {}",
-            missing_flags.join(", ")
-        )));
-    }
-
-    // The negative claim: spellings the parser must not have invented.
-    // Same matcher as `must_contain_flags`, same root-only scope, negated.
-    let present_forbidden: Vec<&str> = contract
-        .must_not_contain_flags
-        .iter()
-        .filter(|spec| flag_present(root, spec))
-        .map(|s| s.as_str())
-        .collect();
-    if !present_forbidden.is_empty() {
-        failures.push(ContractFailure(format!(
-            "must_not_contain_flags: present {}",
-            present_forbidden.join(", ")
-        )));
-    }
-
-    let missing_positionals: Vec<&str> = contract
-        .must_contain_positionals
-        .iter()
-        .filter(|name| {
-            !root
-                .positionals()
-                .any(|p| p.primary_name() == name.as_str())
-        })
-        .map(|s| s.as_str())
-        .collect();
-    if !missing_positionals.is_empty() {
-        failures.push(ContractFailure(format!(
-            "must_contain_positionals: missing {}",
-            missing_positionals.join(", ")
-        )));
-    }
-
-    let missing_modifiers: Vec<&str> = contract
-        .must_contain_modifiers
-        .iter()
-        .filter(|name| !root.modifiers().any(|m| m.primary_name() == name.as_str()))
-        .map(|s| s.as_str())
-        .collect();
-    if !missing_modifiers.is_empty() {
-        failures.push(ContractFailure(format!(
-            "must_contain_modifiers: missing {}",
-            missing_modifiers.join(", ")
-        )));
-    }
-
-    let missing_env_vars: Vec<&str> = contract
-        .must_contain_env_vars
-        .iter()
-        .filter(|name| !root.env_vars().any(|v| v.primary_name() == name.as_str()))
-        .map(|s| s.as_str())
-        .collect();
-    if !missing_env_vars.is_empty() {
-        failures.push(ContractFailure(format!(
-            "must_contain_env_vars: missing {}",
-            missing_env_vars.join(", ")
-        )));
-    }
-
-    for (path, specs) in &contract.must_contain_flags_by_path {
-        let Some(node) = find_node_by_path(root, path) else {
-            failures.push(ContractFailure(format!(
-                "must_contain_flags_by_path: no node at path {path:?}"
-            )));
-            continue;
-        };
-        let missing: Vec<&str> = specs
-            .iter()
-            .filter(|spec| !flag_present(node, spec))
-            .map(|s| s.as_str())
-            .collect();
-        if !missing.is_empty() {
-            failures.push(ContractFailure(format!(
-                "must_contain_flags_by_path[{path:?}]: missing {}",
-                missing.join(", ")
-            )));
-        }
-    }
-
-    failures
-}
-
-/// Resolve a space-separated subcommand path (`"restore"`, `"remote add"`)
-/// against `root`'s own `subcommands`, one path segment per level — the
-/// same walk [`mandible_core::noderef::resolve`] does for the TUI's own
-/// addressing, reimplemented narrowly here rather than pulled in because
-/// this only ever needs a name match, never alias resolution.
-fn find_node_by_path<'a>(root: &'a CommandNode, path: &str) -> Option<&'a CommandNode> {
-    let mut node = root;
-    for segment in path.split_whitespace() {
-        node = node.subcommands.iter().find(|c| c.name == segment)?;
-    }
-    Some(node)
-}
-
-/// Wrap a root already produced by [`extract_tree`] back into an
-/// [`mandible_extract::ExtractionResult`] shape so [`crate::status::compute`]
-/// (which the coverage harness also drives, spec's "one status
-/// definition" requirement) can be reused here without a second
-/// implementation. `tier_statuses`/`tool`/`elapsed` are irrelevant to
-/// `status::compute`, which only ever looks at `root`.
-fn extraction_result_stub(root: CommandNode) -> mandible_extract::ExtractionResult {
-    mandible_extract::ExtractionResult {
-        tool: root.name.clone(),
-        root: Some(root),
-        tier_statuses: Vec::new(),
-        elapsed: Duration::ZERO,
-    }
-}
-
-/// Whether `node`'s own flags satisfy a `must_contain_flags`/
-/// `must_contain_flags_by_path`/`must_not_contain_flags` spec (the last
-/// negated by its caller): `--long-name` matches any
-/// [`mandible_core::Spelling`] with two dashes and that name, `-x` matches
-/// any single-dash single-character spelling, anything else is matched
-/// against every spelling's bare name verbatim. Checks only the one node
-/// given, never recursing.
-///
-/// Checks every spelling, not just `Entity::short`/`Entity::long` (which
-/// return one canonical spelling each): `fold_adjacent_alias_rows` can put
-/// more than one short spelling on an entity (`ffplay`'s `-h, -?, -help,
-/// --help`), so a contract asserting `-?` must still find it once `-h`
-/// claims the canonical slot.
-fn flag_present(node: &CommandNode, spec: &str) -> bool {
-    if let Some(long) = spec.strip_prefix("--") {
-        // Long-*like*, matching `Entity::long`'s own shape rule exactly
-        // (never narrowed to `Dashes::Double` alone): two dashes, or one
-        // dash with more than one character — a single-dash long option
-        // (`ptargrep`'s own `-message`, `Dashes::Single`, four letters)
-        // must still satisfy a `--message` contract entry the way
-        // `Entity::long()` always considered it to.
-        node.flags().any(|f| {
-            f.spellings.iter().any(|s| {
-                (matches!(s.dashes, Dashes::Double)
-                    || (matches!(s.dashes, Dashes::Single) && s.name.chars().count() > 1))
-                    && s.name == long
-            })
-        })
-    } else if let Some(short) = spec.strip_prefix('-') {
-        short.chars().next().is_some_and(|c| {
-            node.flags().any(|f| {
-                f.spellings.iter().any(|s| {
-                    matches!(s.dashes, Dashes::Single)
-                        && s.name.chars().count() == 1
-                        && s.name.starts_with(c)
-                })
-            })
-        })
-    } else {
-        node.flags()
-            .any(|f| f.spellings.iter().any(|s| s.name == spec))
-    }
-}
-
-/// One fixture's outcome, for the runner's summary counters.
-enum Outcome {
-    /// Not marked `[xfail]`, and every check passed.
-    Green,
-    /// Marked `[xfail]`, and at least one check failed as expected.
-    XfailAsExpected,
-    /// A real failure: either a non-`[xfail]` fixture with a failing
-    /// check, a mis-set-up fixture (no snapshot and no `[xfail]`), or a
-    /// strict-xfail violation (an `[xfail]` fixture where every check now
-    /// passes).
-    Failed(String),
-}
-
-/// Run the corpus suite. `bless` rewrites every fixture's `expected.snap`
-/// to match its freshly-extracted tree instead of checking it — blessing
-/// an `[xfail]` fixture is legal (spec's promotion workflow blesses
-/// before removing `[xfail]`; the strict-xfail check on the next plain
-/// run reminds a contributor to remove it).
-///
-/// `format` selects a checking run's report: [`ScoreFormat::Text`] is the
-/// plain per-fixture lines; [`ScoreFormat::Markdown`] additionally builds
-/// a before/after transition table ([`render_markdown_report`]) for
-/// `$GITHUB_STEP_SUMMARY`. Ignored while blessing.
-pub fn run(corpus_root: &Path, bless: bool, format: ScoreFormat) -> anyhow::Result<CorpusReport> {
-    run_with_baseline(corpus_root, bless, format, None)
-}
-
-/// [`run`], additionally comparing every fixture's `[contract]` against
-/// `baseline_root` (a second, plain corpus directory — never git) and
-/// printing a prominent `CONTRACT WEAKENED: ...` line for every field that
-/// got weaker. See [`contract_weakened_lines`] for the full rule and why
-/// this takes a directory instead of a git ref. `None` skips the check
-/// entirely (unchanged behavior from before this existed) — this is what
-/// [`run`] passes, so every existing caller is unaffected.
-pub fn run_with_baseline(
-    corpus_root: &Path,
-    bless: bool,
-    format: ScoreFormat,
-    baseline_root: Option<&Path>,
-) -> anyhow::Result<CorpusReport> {
-    run_with_ceiling(
-        corpus_root,
-        bless,
-        format,
-        MAX_FIXTURE_PARSE_TIME,
-        baseline_root,
-    )
-}
-
-/// [`run`], with the parse-time ceiling injected rather than taken from
-/// [`MAX_FIXTURE_PARSE_TIME`].
-///
-/// Exists so the warn-don't-block property is testable: a real fixture
-/// parses in well under a millisecond, so the only way to exercise a
-/// ceiling violation is to lower the ceiling. Without this seam the
-/// property could only be checked by editing the constant by hand — which
-/// is how it was verified originally, and exactly the kind of unguarded
-/// behaviour a later refactor silently reverses.
-// Ratchet: the fixture runner's main loop; split into load and compare is tracked work. Listed in scripts/ratchet.txt.
-#[allow(clippy::too_many_lines)]
-fn run_with_ceiling(
-    corpus_root: &Path,
-    bless: bool,
-    format: ScoreFormat,
-    max_parse_time: Duration,
-    baseline_root: Option<&Path>,
-) -> anyhow::Result<CorpusReport> {
-    let fixtures = discover_fixtures(corpus_root)?;
-    if fixtures.is_empty() {
-        anyhow::bail!(
-            "no fixtures found under {} (expected corpus/<tool>/<version>/meta.toml)",
-            corpus_root.display()
-        );
-    }
-    // Contract-weakening lines (see `contract_weakened_lines`'s doc
-    // comment) — computed once, up front, from plain `meta.toml` reads
-    // only (no extraction, no git), so a failure here (a malformed
-    // baseline directory) surfaces before any of the real, expensive work
-    // below starts.
-    let weakened: Vec<String> = match baseline_root {
-        Some(dir) => {
-            let baseline_fixtures = discover_fixtures(dir)?;
-            contract_weakened_lines(&fixtures, &baseline_fixtures)
-        }
-        None => Vec::new(),
-    };
-
-    let mut lines = Vec::new();
-    let mut outcomes = Vec::new();
-    let mut timing_violations = Vec::new();
-    // Only populated for a checking run (`bless: false`) — see
-    // `render_markdown_report`. Cheap to build unconditionally alongside
-    // `lines`: every fixture here is a few KiB, never worth gating on.
-    let mut fixture_rows: Vec<FixtureRow> = Vec::new();
-
-    for fixture in &fixtures {
-        let transcript = fixture.build_transcript()?;
-        let runner = Runner::new(default_tiers_with_probe(Arc::new(transcript)));
-        let resolved = fixture.resolved_tool();
-
-        let start = Instant::now();
-        let root = extract_tree(&runner, &resolved);
-        let elapsed = start.elapsed();
-
-        if elapsed > max_parse_time {
-            timing_violations.push(format!(
-                "{}: parsed in {:?}, exceeding the {:?} ceiling",
-                fixture.label, elapsed, max_parse_time
-            ));
-        }
-
-        if bless {
-            match &root {
-                Some(root) => {
-                    let rendered = render_snapshot(root)?;
-                    std::fs::write(fixture.expected_snap_path(), &rendered).map_err(|e| {
-                        anyhow::anyhow!("writing {}: {e}", fixture.expected_snap_path().display())
-                    })?;
-                    lines.push(format!("blessed {} ({:?})", fixture.label, elapsed));
-                }
-                None => {
-                    lines.push(format!(
-                        "{}: no root produced — nothing to bless",
-                        fixture.label
-                    ));
-                }
-            }
-            continue;
-        }
-
-        let snapshot_check = check_snapshot(fixture, root.as_ref())?;
-        let contract_failures = check_contract(&fixture.meta.contract, root.as_ref());
-        let is_xfail = fixture.meta.xfail.as_ref().is_some_and(|x| x.broken);
-
-        // A missing `expected.snap` is legal only for a fixture still
-        // marked `[xfail]` (corpus/README.md step 4: "a fixture marked
-        // broken has no expected tree yet"). For anything else it's a
-        // real gap — a "green" fixture with no pinned tree asserts
-        // nothing about structure at all, defeating the ratchet.
-        let snapshot_ok = match &snapshot_check {
-            SnapshotCheck::Match => true,
-            SnapshotCheck::Missing => is_xfail,
-            SnapshotCheck::NoRoot | SnapshotCheck::Mismatch { .. } => false,
-        };
-        let all_pass = snapshot_ok && contract_failures.is_empty();
-
-        let mut detail = Vec::new();
-        for failure in &contract_failures {
-            detail.push(format!("contract: {}", failure.0));
-        }
-        match &snapshot_check {
-            SnapshotCheck::Match => detail.push("snapshot: match".to_string()),
-            SnapshotCheck::Missing if is_xfail => {
-                detail.push("snapshot: none yet (legal while [xfail])".to_string())
-            }
-            SnapshotCheck::Missing => detail.push(
-                "snapshot: missing expected.snap (required unless marked [xfail])".to_string(),
-            ),
-            SnapshotCheck::NoRoot => {
-                detail.push("snapshot: no root produced by any tier".to_string())
-            }
-            SnapshotCheck::Mismatch {
-                line,
-                expected,
-                actual,
-            } => detail.push(format!(
-                "snapshot mismatch at expected.snap:{line}: expected `{expected}` got `{actual}`"
-            )),
-        }
-
-        // Only when a scope is actually recorded — an unscoped fixture
-        // (the overwhelming majority, today) stays silent here exactly
-        // as it always has; `show_fixture` is where "unscoped" is spelled
-        // out explicitly for a fixture being inspected one at a time.
-        // Purely informational: never affects `all_pass`, since a scope
-        // claim is a record of what a human checked, not itself a check.
-        if !fixture.meta.contract.verdict_scope.is_empty() {
-            detail.push(format!(
-                "verdict_scope: {}",
-                verdict_scope_label(&fixture.meta.contract.verdict_scope)
-            ));
-        }
-
-        // Unlike `verdict_scope` above, `provenance` is always set (it is
-        // a required field — `discover_fixtures`'s guard), so this note
-        // always prints; there is no "unscoped"-style silent case here.
-        detail.push(format!(
-            "provenance: {}",
-            provenance_label(fixture.meta.bless.provenance)
-        ));
-
-        let outcome = if is_xfail {
-            if all_pass {
-                // The promote message belongs in `detail` too, not just
-                // the returned `Outcome::Failed` message — otherwise it
-                // only ever reaches `CorpusReport::failures` and never
-                // the per-fixture line this loop prints, which is
-                // supposed to be the "name the tool and what broke" text
-                // a human actually reads.
-                detail.push(
-                    "[xfail] but every check now passes — the bug appears fixed; promote it \
-                     (remove [xfail], commit expected.snap if it isn't already)"
-                        .to_string(),
-                );
-                Outcome::Failed(format!("{}: {}", fixture.label, detail.join("; ")))
-            } else {
-                Outcome::XfailAsExpected
-            }
-        } else if all_pass {
-            Outcome::Green
-        } else {
-            Outcome::Failed(format!("{}: {}", fixture.label, detail.join("; ")))
-        };
-
-        let status_word = match &outcome {
-            Outcome::Green => "ok",
-            Outcome::XfailAsExpected => "xfail (as expected)",
-            Outcome::Failed(_) => "FAIL",
-        };
-        lines.push(format!(
-            "{:<24} {:<20} ({:?})  {}",
-            fixture.label,
-            status_word,
-            elapsed,
-            detail.join("; ")
-        ));
-
-        fixture_rows.push(FixtureRow {
-            label: fixture.label.clone(),
-            status_word,
-            detail: detail.clone(),
-            current: summarize(&fixture.meta.tool.name, root.as_ref()),
-            previous: previous_summary(fixture),
-            verdict_scope: fixture.meta.contract.verdict_scope.clone(),
-            provenance: fixture.meta.bless.provenance,
-        });
-
-        outcomes.push(outcome);
-    }
-
-    let failed_count = outcomes
-        .iter()
-        .filter(|o| matches!(o, Outcome::Failed(_)))
-        .count();
-    let failures: Vec<String> = outcomes
-        .iter()
-        .filter_map(|o| match o {
-            Outcome::Failed(msg) => Some(msg.clone()),
-            _ => None,
-        })
-        .collect();
-    // Deliberately NOT extended with `timing_violations` — the parse-time
-    // ceiling warns, it does not fail the run. See
-    // [`MAX_FIXTURE_PARSE_TIME`] for why.
-
-    let green = outcomes
-        .iter()
-        .filter(|o| matches!(o, Outcome::Green))
-        .count();
-    let xfail = outcomes
-        .iter()
-        .filter(|o| matches!(o, Outcome::XfailAsExpected))
-        .count();
-
-    lines.push(String::new());
-    if bless {
-        lines.push(format!("blessed {} fixture(s)", fixtures.len()));
-    } else {
-        let ok_provenance = provenance_split_label(provenance_counts(
-            fixture_rows
-                .iter()
-                .filter(|r| r.status_word == "ok")
-                .map(|r| r.provenance),
-        ));
-        lines.push(format!(
-            "{} fixture(s): {green} ok ({ok_provenance}), {xfail} xfail (as expected), \
-             {failed_count} failed",
-            fixtures.len(),
-        ));
-        for violation in &timing_violations {
-            // "warning", not a bare label: this line does not fail the
-            // run, and a reader who cannot tell that from the output will
-            // either chase a non-failure or learn to skim past real ones.
-            lines.push(format!("warning: slow parse (does not fail): {violation}"));
-        }
-    }
-
-    // Markdown only applies to a checking run — see `run`'s doc comment.
-    // It fully replaces `lines.join(...)` rather than appending to it: the
-    // markdown report is a self-contained artifact (table plus `<details>`
-    // for anything capped), not the plain-text report with formatting
-    // bolted on.
-    let text = if !bless && matches!(format, ScoreFormat::Markdown) {
-        render_markdown_report(
-            &fixture_rows,
-            fixtures.len(),
-            green,
-            xfail,
-            failed_count,
-            &weakened,
-        )
-    } else {
-        // Contract-weakening lines go first, ahead of every per-fixture
-        // line and the summary — "prominent... so a reviewer skimming a
-        // green run cannot miss it" means the very top of the output, not
-        // one more line in a footer nobody reads once green.
-        let mut prefixed = weakened;
-        prefixed.extend(lines);
-        prefixed.join("\n")
-    };
-
-    Ok(CorpusReport {
-        text,
-        failures,
-        bless,
-    })
-}
-
-/// One side of a fixture's semantic comparison (spec companion work
-/// order's requirement: a report must say *what changed*, never present a
-/// text diff) — either the tree `expected.snap` currently pins, or the
-/// tree this run just extracted. Built once via [`summarize`] so both
-/// sides go through identical logic; no separate "previous" rules to drift
-/// from "current" ones.
-struct TreeSummary {
-    /// [`crate::status::compute`]'s label — the *same* function
-    /// `check_contract`'s `min_status` check uses, so a status shown here
-    /// can never disagree with what actually gated the run (status.rs's
-    /// own doc comment: "two independent definitions of 'status' will
-    /// drift, and the drift will be discovered at the worst possible
-    /// time").
-    status: &'static str,
-    nodes: usize,
-    flags: usize,
-    /// Every subcommand's dotted path (`"bisect start"`, not just
-    /// `"start"`), so two same-named subcommands under different parents
-    /// are never conflated.
-    subcommands: BTreeSet<String>,
-    /// Every flag's canonical spelling (`--long` if present, else `-x`).
-    flag_names: BTreeSet<String>,
-}
-
-/// Build a [`TreeSummary`] for `root` (`None` when no tier produced
-/// anything, e.g. a "current" side with no root, or an unresolvable
-/// `expected.snap`). `tool` only feeds the status stub's `tool` field,
-/// which `crate::status::compute` never actually reads.
-fn summarize(tool: &str, root: Option<&CommandNode>) -> TreeSummary {
-    let stub = mandible_extract::ExtractionResult {
-        tool: tool.to_string(),
-        root: root.cloned(),
-        tier_statuses: Vec::new(),
-        elapsed: Duration::ZERO,
-    };
-    let status = crate::status::compute(&stub);
-    let mut subcommands = BTreeSet::new();
-    let mut flag_names = BTreeSet::new();
-    if let Some(r) = root {
-        collect_subcommand_paths(r, "", &mut subcommands);
-        collect_flag_names(r, &mut flag_names);
-    }
-    TreeSummary {
-        status: status.label,
-        nodes: root.map(count_nodes).unwrap_or(0),
-        flags: root.map(count_flags).unwrap_or(0),
-        subcommands,
-        flag_names,
-    }
-}
-
-fn count_nodes(node: &CommandNode) -> usize {
-    1 + node.subcommands.iter().map(count_nodes).sum::<usize>()
-}
-
-fn count_flags(node: &CommandNode) -> usize {
-    node.flags().count() + node.subcommands.iter().map(count_flags).sum::<usize>()
-}
-
-fn collect_flag_names(node: &CommandNode, out: &mut BTreeSet<String>) {
-    for f in node.flags() {
-        if let Some(long) = f.long() {
-            out.insert(format!("--{long}"));
-        } else if let Some(short) = f.short() {
-            out.insert(format!("-{short}"));
-        }
-    }
-    for child in &node.subcommands {
-        collect_flag_names(child, out);
-    }
-}
-
-fn collect_subcommand_paths(node: &CommandNode, prefix: &str, out: &mut BTreeSet<String>) {
-    for child in &node.subcommands {
-        let path = if prefix.is_empty() {
-            child.name.clone()
-        } else {
-            format!("{prefix} {}", child.name)
-        };
-        out.insert(path.clone());
-        collect_subcommand_paths(child, &path, out);
-    }
-}
-
-/// Minimal, tolerant mirror of `expected.snap`'s compact YAML shape
-/// (`mandible_core::snapshot::NodeSnapshot`), carrying only what
-/// [`TreeSummary`] needs. `NodeSnapshot` itself derives only `Serialize`
-/// (its own doc comment: the compact form is a one-way review artifact,
-/// not a round-trip format — every empty `Vec`/`None` field is omitted
-/// entirely), so a direct `Deserialize` of the full IR would fail on
-/// "missing field" for nearly every real fixture. `#[serde(default)]` on
-/// every field here is what makes an omitted key deserialize back to the
-/// same "empty" value its serializer chose not to write, instead of an
-/// error.
-#[derive(Debug, Clone, Default, Deserialize)]
-struct SnapFlag {
-    /// Every rendered spelling, in document order — `expected.snap`'s
-    /// `spellings` key, exactly as [`mandible_core::FlagSnapshot`] writes
-    /// it. Parsed back into [`Spelling`]s by [`parse_rendered_spelling`]
-    /// so [`collect_flag_names`] can read `short()`/`long()` off the
-    /// reconstructed entity the same way it reads them off a fresh one.
-    #[serde(default)]
-    spellings: Vec<String>,
-    /// Needed for nothing this module names directly, but
-    /// [`crate::status::compute`]'s `pct_flags_with_text` (and therefore the
-    /// `low-confidence` vs `ok` status boundary) reads it — omitting it
-    /// would make every reconstructed "previous" tree look 0% described
-    /// regardless of the real fixture, which would desync `TreeSummary`'s
-    /// two sides even when `expected.snap` and the fresh extraction are
-    /// byte-identical.
-    #[serde(default)]
-    description: Option<String>,
-}
-
-#[derive(Debug, Clone, Default, Deserialize)]
-struct SnapNode {
-    #[serde(default)]
-    name: String,
-    #[serde(default)]
-    summary: Option<String>,
-    #[serde(default)]
-    flags: Vec<SnapFlag>,
-    #[serde(default)]
-    subcommands: Vec<SnapNode>,
-    #[serde(default)]
-    unparsed: Vec<String>,
-    #[serde(default)]
-    heading_attested: bool,
-    #[serde(default)]
-    invocation_attested: bool,
-}
-
-/// Rebuild a real (if synthetic-provenance) [`CommandNode`] from a
-/// [`SnapNode`], so [`summarize`] can run the *exact same* status/count/
-/// name-collection logic over `expected.snap`'s tree that it runs over a
-/// freshly extracted one — one code path for both sides of the
-/// comparison, never a second one hand-written against the compact YAML
-/// shape. Every field `SnapNode` doesn't carry (provenance detail,
-/// descriptions, positionals, ...) is irrelevant to what [`summarize`]
-/// reads and is left at `CommandNode::new`'s defaults.
-fn snap_to_command_node(n: &SnapNode) -> CommandNode {
-    let mut node = CommandNode::new(n.name.clone(), Provenance::single(Source::HelpText));
-    node.summary = n.summary.as_deref().map(Text::sanitize);
-    node.heading_attested = n.heading_attested;
-    node.invocation_attested = n.invocation_attested;
-    node.unparsed = n.unparsed.iter().map(|s| Text::sanitize(s)).collect();
-    node.set_flags(
-        n.flags
-            .iter()
-            .map(|f| {
-                let mut flag = Entity::new(EntityKind::Flag, Provenance::single(Source::HelpText));
-                flag.spellings = f
-                    .spellings
-                    .iter()
-                    .map(|s| parse_rendered_spelling(s))
-                    .collect();
-                flag.description = f.description.as_deref().map(Text::sanitize);
-                flag
-            })
-            .collect(),
-    );
-    node.subcommands = n.subcommands.iter().map(snap_to_command_node).collect();
-    node
-}
-
-/// Parse one of [`FlagSnapshot`](mandible_core::FlagSnapshot)'s rendered
-/// `spellings` entries (`"-i"`, `"--interactive"`, `"--[no-]color"`,
-/// `"-help"`, `"-r[esolve]"`) back into a [`Spelling`], the inverse of
-/// [`Spelling::render`]. Good enough for [`snap_to_command_node`]'s
-/// purpose — reconstructing an entity whose `short()`/`long()` accessors
-/// agree with what a fresh extraction would produce — not a general
-/// parser: a name that itself began with a literal `-` would round-trip
-/// wrong, and no real flag spelling does.
-fn parse_rendered_spelling(rendered: &str) -> Spelling {
-    let (dashes, rest) = if let Some(r) = rendered.strip_prefix("--") {
-        (Dashes::Double, r)
-    } else if let Some(r) = rendered.strip_prefix('-') {
-        (Dashes::Single, r)
-    } else {
-        (Dashes::None, rendered)
-    };
-    if let Some(base) = rest.strip_prefix("[no-]") {
-        return Spelling {
-            name: base.to_string(),
-            dashes,
-            negatable: true,
-            abbrev: None,
-        };
-    }
-    // An abbreviation bracket: "r[esolve]" -> name "resolve", abbrev
-    // Some(1) — the prefix's character count.
-    if let Some(open) = rest.find('[') {
-        if let Some(rest_after) = rest.strip_suffix(']') {
-            let (prefix, bracketed) = rest_after.split_at(open);
-            let bracketed = &bracketed[1..]; // drop the '['
-            return Spelling {
-                name: format!("{prefix}{bracketed}"),
-                dashes,
-                negatable: false,
-                abbrev: Some(prefix.chars().count()),
-            };
-        }
-    }
-    Spelling {
-        name: rest.to_string(),
-        dashes,
-        negatable: false,
-        abbrev: None,
-    }
-}
-
-/// Read and convert a fixture's `expected.snap` into a [`TreeSummary`],
-/// `None` when it doesn't exist yet (legal for an unfixed `[xfail]`
-/// fixture, `corpus/README.md` step 4) or fails to parse (a corrupt or
-/// hand-edited file — degrade to "no baseline" rather than panicking on
-/// tool input, per AGENTS.md's `unwrap()` rule).
-fn previous_summary(fixture: &Fixture) -> Option<TreeSummary> {
-    let snap_path = fixture.expected_snap_path();
-    if !snap_path.is_file() {
-        return None;
-    }
-    let raw = std::fs::read_to_string(&snap_path).ok()?;
-    let snap: SnapNode = serde_yaml::from_str(&raw).ok()?;
-    let converted = snap_to_command_node(&snap);
-    Some(summarize(&fixture.meta.tool.name, Some(&converted)))
-}
-
-/// One fixture's row in the markdown transition report: the same
-/// pass/fail classification and detail lines the text report already
-/// computed, plus the two [`TreeSummary`] sides [`render_markdown_report`]
-/// diffs.
-struct FixtureRow {
-    label: String,
-    status_word: &'static str,
-    /// `contract: ...` / `snapshot: ...` lines, reused to decide which
-    /// remedy to name (§ this module's doc comment on `render_markdown_report`).
-    detail: Vec<String>,
-    current: TreeSummary,
-    previous: Option<TreeSummary>,
-    /// What a human actually verified before this fixture was blessed
-    /// (`ContractMeta::verdict_scope`'s doc comment) — surfaced as the
-    /// table's own `scope` column so a reviewer sees, without opening
-    /// `meta.toml`, that a green row's descriptions may still be
-    /// unreviewed prose.
-    verdict_scope: Vec<VerdictScope>,
-    /// Who blessed this fixture's `expected.snap` (`BlessMeta`'s doc
-    /// comment) — the complement to `verdict_scope`: surfaced as its own
-    /// table column so an `ok` row can never be misread as human-verified
-    /// when no human has looked.
-    provenance: BlessProvenance,
-}
-
-/// Cap on how many names a single markdown table cell shows inline before
-/// folding the rest into "+N more" — keeps a large-fixture regression
-/// scannable. Not load-bearing (informational only): the complete list is
-/// never dropped, only moved into a `<details>` block below the table (see
-/// [`capped_names`]), matching `path-sweep.yml`'s own summary step
-/// (`<details><summary>Per-tool tables</summary>`) for exactly the same
-/// reason — GitHub's step-summary UI renders `<details>` natively, so a
-/// reviewer who wants the full list expands it without ever leaving the
-/// page, and one who doesn't isn't shown a 1,000-line dump by default.
-const MARKDOWN_NAME_CAP: usize = 8;
-
-/// Render `names` as a capped, comma-joined inline string, plus (only when
-/// truncated) the complete list for a `<details>` block.
-fn capped_names(names: &BTreeSet<String>) -> (String, Option<String>) {
-    if names.is_empty() {
-        return (String::new(), None);
-    }
-    let all: Vec<&str> = names.iter().map(String::as_str).collect();
-    if all.len() <= MARKDOWN_NAME_CAP {
-        return (all.join(", "), None);
-    }
-    let shown = all[..MARKDOWN_NAME_CAP].join(", ");
-    let capped = format!("{shown}, +{} more", all.len() - MARKDOWN_NAME_CAP);
-    (capped, Some(all.join(", ")))
-}
-
-/// Escape the one GFM table-breaking character. Same reasoning and same
-/// narrow scope as `coverage::md_escape` — tool/flag/subcommand names are
-/// the only free-form content a table cell here ever holds.
-fn md_escape(s: &str) -> String {
-    s.replace('|', "\\|")
-}
-
-/// Render the corpus run as a GitHub-flavored markdown transition report
-/// for `$GITHUB_STEP_SUMMARY` — never a text diff of `expected.snap`
-/// (a 1,000+ line YAML diff is unreviewable and teaches reviewers to
-/// approve unread). Reports what changed semantically instead: status,
-/// node/flag counts, and which named subcommands/flags appeared or
-/// disappeared.
-///
-/// Passing fixtures are included too, compactly ("no change") — an
-/// all-failures report gives no sense of what the ratchet protects.
-///
-/// The remedy named for a failing row depends on which check failed: a
-/// snapshot mismatch names `--bless`; a `[contract]` violation names a
-/// deliberate `meta.toml` edit (the contract may only weaken by an
-/// explicit change, never by blessing).
-fn render_markdown_report(
-    rows: &[FixtureRow],
-    total: usize,
-    green: usize,
-    xfail: usize,
-    failed: usize,
-    weakened: &[String],
-) -> String {
-    let mut out = String::new();
-    // Contract weakening, ahead of the report's own heading — legal per
-    // corpus/README.md but must never be quiet. `> [!WARNING]` renders
-    // with its own colored icon on GitHub, unlike a plain bullet.
-    if !weakened.is_empty() {
-        out.push_str("> [!WARNING]\n");
-        for line in weakened {
-            out.push_str(&format!("> {}\n", md_escape(line)));
-        }
-        out.push('\n');
-    }
-    out.push_str("## Corpus regression report\n\n");
-    out.push_str(
-        "Every fixture below is replayed through the real tiered extraction pipeline from \
-         frozen bytes — zero subprocesses, nothing environment-dependent, no live tool version \
-         to drift. A red row is always a real parser change, never runner flakiness \
-         (`corpus/README.md`); that is exactly why this check is a hard gate rather than a \
-         reported-only sweep like the PATH sweep.\n\n",
-    );
-    let ok_provenance = provenance_split_label(provenance_counts(
-        rows.iter()
-            .filter(|r| r.status_word == "ok")
-            .map(|r| r.provenance),
-    ));
-    out.push_str(&format!(
-        "**{total} fixture(s):** {green} ok ({ok_provenance}), {xfail} xfail (as expected), \
-         {failed} failed.\n\n",
-    ));
-    out.push_str("| fixture | outcome | status | nodes | flags | scope | provenance | change |\n");
-    out.push_str("|---|---|---|---|---|---|---|---|\n");
-
-    let mut details_sections: Vec<String> = Vec::new();
-
-    for row in rows {
-        let (status_cell, nodes_cell, flags_cell, mut change_parts) = match &row.previous {
-            None => (
-                format!("{} (no baseline)", row.current.status),
-                row.current.nodes.to_string(),
-                row.current.flags.to_string(),
-                vec!["no `expected.snap` yet".to_string()],
-            ),
-            Some(prev) => {
-                let status_cell = if prev.status == row.current.status {
-                    prev.status.to_string()
-                } else {
-                    format!("{} → {}", prev.status, row.current.status)
-                };
-                let nodes_cell = if prev.nodes == row.current.nodes {
-                    prev.nodes.to_string()
-                } else {
-                    format!("{}→{}", prev.nodes, row.current.nodes)
-                };
-                let flags_cell = if prev.flags == row.current.flags {
-                    prev.flags.to_string()
-                } else {
-                    format!("{}→{}", prev.flags, row.current.flags)
-                };
-
-                let mut parts = Vec::new();
-                let removed_subs: BTreeSet<String> = prev
-                    .subcommands
-                    .difference(&row.current.subcommands)
-                    .cloned()
-                    .collect();
-                let added_subs: BTreeSet<String> = row
-                    .current
-                    .subcommands
-                    .difference(&prev.subcommands)
-                    .cloned()
-                    .collect();
-                let removed_flags: BTreeSet<String> = prev
-                    .flag_names
-                    .difference(&row.current.flag_names)
-                    .cloned()
-                    .collect();
-                let added_flags: BTreeSet<String> = row
-                    .current
-                    .flag_names
-                    .difference(&prev.flag_names)
-                    .cloned()
-                    .collect();
-
-                for (kind, names) in [
-                    ("removed subcommands", &removed_subs),
-                    ("added subcommands", &added_subs),
-                    ("removed flags", &removed_flags),
-                    ("added flags", &added_flags),
-                ] {
-                    if names.is_empty() {
-                        continue;
-                    }
-                    let count = names.len();
-                    let (capped, full) = capped_names(names);
-                    parts.push(format!("{kind} ({count}): {capped}"));
-                    if let Some(full) = full {
-                        details_sections.push(format!(
-                            "<details><summary>{} — {count} {kind}</summary>\n\n{full}\n\n</details>",
-                            row.label,
-                        ));
-                    }
-                }
-                if parts.is_empty() {
-                    // A failing fixture whose tracked dimensions all match
-                    // still has a real snapshot mismatch somewhere the
-                    // summary cannot see: confidence, provenance, usage
-                    // lines, a reworded description. Printing "no change"
-                    // beside FAIL reads as a bug in the runner rather than
-                    // a regression in the parser, and sends the reviewer
-                    // hunting in the wrong place. Measured: scaling
-                    // `compute_confidence` by 0.8 produced exactly that
-                    // row in a real CI run.
-                    parts.push(
-                        if row.status_word == "FAIL" {
-                            "differs outside status/nodes/flags — check confidence, \
-                             usage or descriptions in the snapshot diff"
-                        } else {
-                            "no change"
-                        }
-                        .to_string(),
-                    );
-                }
-                (status_cell, nodes_cell, flags_cell, parts)
-            }
-        };
-
-        if row.status_word == "FAIL" {
-            let has_snapshot_mismatch = row
-                .detail
-                .iter()
-                .any(|d| d.starts_with("snapshot mismatch") || d.contains("missing expected.snap"));
-            let has_contract_failure = row.detail.iter().any(|d| d.starts_with("contract:"));
-            if has_snapshot_mismatch {
-                change_parts.push(
-                    "**fix:** `cargo run -p xtask -- corpus --bless`, then review the \
-                     `expected.snap` diff"
-                        .to_string(),
-                );
-            }
-            if has_contract_failure {
-                change_parts.push(
-                    "**fix:** `[contract]` is normative — edit `meta.toml` deliberately \
-                     (never `--bless`) and justify the change in the PR"
-                        .to_string(),
-                );
-            }
-        }
-
-        out.push_str(&format!(
-            "| {} | {} | {} | {} | {} | {} | {} | {} |\n",
-            md_escape(&row.label),
-            row.status_word,
-            md_escape(&status_cell),
-            nodes_cell,
-            flags_cell,
-            md_escape(&verdict_scope_label(&row.verdict_scope)),
-            provenance_label(row.provenance),
-            md_escape(&change_parts.join("; ")),
-        ));
-    }
-
-    if !details_sections.is_empty() {
-        out.push('\n');
-        for section in &details_sections {
-            out.push_str(section);
-            out.push('\n');
-        }
-    }
-
-    out
-}
-
-/// The outcome of a full corpus run.
-pub struct CorpusReport {
-    /// Human-readable per-fixture results plus a summary line.
-    pub text: String,
-    /// Every reason the run should fail (contract/snapshot/strict-xfail
-    /// violations, plus any parse-time-ceiling violations), empty when
-    /// everything is clean. Always empty in `--bless` mode.
-    pub failures: Vec<String>,
-    bless: bool,
-}
-
-impl CorpusReport {
-    /// True when this run should exit non-zero.
-    pub fn failed(&self) -> bool {
-        !self.bless && !self.failures.is_empty()
-    }
-}
-
-/// Print one fixture's captured help text beside the tree the parser makes
-/// of it, then return. Renders the same side-by-side comparison `xtask
-/// audit emit` produces for a live tool, sourced from the frozen capture,
-/// so an `[xfail]` fixture's asserted defect can be seen directly.
-///
-/// Read-only and separate from the checking run: neither blesses nor fails.
-/// One fixture replayed through the real pipeline: the raw help text the
-/// tiers built from, and the tree they produced.
-pub struct ReplayedFixture {
-    /// The fixture's tool name (`meta.tool.name`), which is what an audit
-    /// entry is keyed on.
-    pub tool: String,
-    /// The raw help text, chosen by the same expansion/`--help`/`-h` rule
-    /// the live oracles use ([`crate::misattribution::root_help_text_from`])
-    /// so a fixture-sourced detector run and a sweep-sourced one are reading
-    /// the same bytes for the same tool.
-    pub raw: String,
-    /// The extracted root, or `None` when no tier produced one.
-    pub root: Option<CommandNode>,
-}
-
-/// Replay every fixture whose directory name is `version` (e.g.
-/// `audit-seed2`) and hand back what each one parsed to.
-///
-/// Zero subprocesses, exactly like [`run`]: this is the same frozen-bytes
-/// replay the corpus suite performs, exposed so `crate::detector` can run a
-/// detector over the audited tools without a `PATH` sweep. Fixtures that
-/// carry no usable help capture are skipped rather than yielded with an
-/// empty `raw`, so a caller cannot mistake "nothing was captured" for "the
-/// tool's help text is empty".
-pub fn replay_version(corpus_root: &Path, version: &str) -> anyhow::Result<Vec<ReplayedFixture>> {
-    let mut out = Vec::new();
-    for fixture in discover_fixtures(corpus_root)? {
-        if !fixture.label.ends_with(&format!("/{version}")) {
-            continue;
-        }
-        let transcript = fixture.build_transcript()?;
-        let mut recordings = HashMap::new();
-        for capture in &fixture.meta.captures {
-            let key = capture.argv[1..].to_vec();
-            recordings.insert(
-                key,
-                ExecOutput {
-                    stdout: read_capture_file(&fixture.dir, &capture.stdout)?,
-                    stderr: match &capture.stderr {
-                        Some(name) => read_capture_file(&fixture.dir, name)?,
-                        None => Vec::new(),
-                    },
-                    exit_code: Some(capture.exit_code.unwrap_or(0)),
-                    timed_out: false,
-                },
-            );
-        }
-        let Some(raw) = crate::misattribution::root_help_text_from(&recordings) else {
-            continue;
-        };
-        let runner = Runner::new(default_tiers_with_probe(Arc::new(transcript)));
-        let root = extract_tree(&runner, &fixture.resolved_tool());
-        out.push(ReplayedFixture {
-            tool: fixture.meta.tool.name.clone(),
-            raw,
-            root,
-        });
-    }
-    Ok(out)
-}
-
-pub fn show_fixture(corpus_root: &Path, pattern: &str) -> anyhow::Result<()> {
-    let fixtures = discover_fixtures(corpus_root)?;
-    let matches: Vec<&Fixture> = fixtures
-        .iter()
-        .filter(|f| f.label.contains(pattern))
-        .collect();
-
-    let fixture = match matches.as_slice() {
-        [] => anyhow::bail!(
-            "no fixture matching {pattern:?} under {}. Available: {}",
-            corpus_root.display(),
-            fixtures
-                .iter()
-                .map(|f| f.label.as_str())
-                .collect::<Vec<_>>()
-                .join(", ")
-        ),
-        [one] => *one,
-        many => anyhow::bail!(
-            "{pattern:?} matches {} fixtures: {}. Narrow it.",
-            many.len(),
-            many.iter()
-                .map(|f| f.label.as_str())
-                .collect::<Vec<_>>()
-                .join(", ")
-        ),
-    };
-
-    println!("fixture: {}", fixture.label);
-    println!("path:    {}", fixture.dir.display());
-    if let Some(xfail) = &fixture.meta.xfail {
-        println!(
-            "status:  [xfail] {}",
-            xfail.reason.as_deref().unwrap_or("(no reason recorded)")
-        );
-    } else {
-        println!("status:  expected to pass");
-    }
-    if fixture.meta.contract.verdict_scope.is_empty() {
-        println!(
-            "scope:   unscoped — no dimension of this tree is asserted human-verified \
-             (a passing snapshot check still freezes every field, descriptions included)"
-        );
-    } else {
-        println!(
-            "scope:   {} — only these dimensions were human-verified before this fixture was \
-             blessed; the rest of the tree is frozen but unreviewed",
-            verdict_scope_label(&fixture.meta.contract.verdict_scope)
-        );
-    }
-    println!(
-        "provenance: {} — who blessed this fixture's expected.snap (see corpus/README.md)",
-        provenance_label(fixture.meta.bless.provenance)
-    );
-    println!();
-
-    for capture in &fixture.meta.captures {
-        let argv = capture.argv.join(" ");
-        let files: [(&str, Option<&str>); 2] = [
-            ("stdout", Some(capture.stdout.as_str())),
-            ("stderr", capture.stderr.as_deref()),
-        ];
-        for (label, file) in files {
-            let Some(name) = file else { continue };
-            let bytes = std::fs::read(fixture.dir.join(name))?;
-            if bytes.is_empty() {
-                continue;
-            }
-            println!("=== captured: {argv}  ({label}) ===");
-            println!("{}", String::from_utf8_lossy(&bytes));
-        }
-    }
-
-    let transcript = fixture.build_transcript()?;
-    let runner = Runner::new(default_tiers_with_probe(Arc::new(transcript)));
-    let resolved = fixture.resolved_tool();
-    println!("=== parsed tree ===");
-    match extract_tree(&runner, &resolved) {
-        Some(node) => println!("{}", render_snapshot(&node)?),
-        None => println!("(no tier produced a root node)"),
-    }
-    Ok(())
 }
 
 #[cfg(test)]
